Bug 33178: Use template wrapper for authority and bibliographic subfield entry form tabs
This patch updates the authority and bibliographic subfield entry
forms to replace tab markup with the use of WRAPPERs. Some JS has been
modified in order to adapt to the changed Bootstrap tab markup.

The patch also updates the WRAPPER code so that a unique ID can be
added to individual tabs. This is necessary as a hook for the Sortable
JavaScript.

To test, apply the patch and go to Administration.

- Go to MARC bibliographic frameworks -> Actions -> MARC structure.
  - Click Actions -> View subfields for a tag with more than one
    subfield.
    - Click the "Edit" button for one of the subfields, not the first
      one in the list.
      - The tab for the subfield you chose should be active.
      - Confirm that all the tabs look correct and work correctly.
      - Try dragging one of the tabs into a new position. Save your
        changes and confirm that the subfield was saved in that new
        position.
- Go to Administration -> Authority types -> Actions -> MARC structure
  and repeat the same steps above with the authority record subfields.

Sponsored-By: Athens County Public Libraries
Signed-off-by: Phil Ringnalda <phil@chetcolibrary.org>
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>