Bug 14787: Fix tidying issue in circulation.pl
There are large blocks of code in circulation.pl that are incorrectly indented. This means any new code placed within those blocks fails the QA tests as it automatically adds new untidy lines. This patch tidies the file to remove this issue and start from a clean position in terms of tidyness.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>