Bug 7508: Collapsable items on items tab
Since we recently updated the patron entry form in the staff interface
to have more visibly-collapsible sections I thought there might be
interest in picking up this old bug.

The patch implements collapsible sections on the item detail page,
allowing the user to hide sections by clicking the item barcode header.
The patch includes relocation of some JS code from the authority and
MARC editors into the global JS file. This is the code that helps jump
the user to a certain scroll point on the page.

To test, apply the patch and clear your browser cache.

- In the staff interface, locate a bibliographic record with items and
  view the detail page.
- Click the "Items" tab in the sidebar.
- On the item detail page, each section should have an icon indicating
  that they are expanded. Hovering your mouse over the section heading
  should highlight the clickable area.
- Test that each section collapses correctly and that the next section
  is moved into view.

- Go to Cataloging -> New record. Test that the links to specific tags
  under the main tab bar still work to jump you to the correct part of
  the page.
- Go to Authorities -> New authority and perform the same test.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Brendan Lawlor <blawlor@clamsnet.org>
Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>